requirements.txt
appdirs==1.4.3
autopep8==1.4.4
CacheControl==0.12.6
cairocffi==1.1.0
CairoSVG==2.4.2
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
cssselect2==0.2.2
defusedxml==0.6.0
distlib==0.3.0
distro==1.4.0
dj-database-url==0.5.0
Django==2.2.7
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap-form==3.4
django-cors-headers==3.3.0
django-filter==2.2.0
django-mysql==3.2.0
django-tempus-dominus==5.1.2.9
djangorestframework==3.10.3
et-xmlfile==1.0.1
html5lib==1.0.1
idna==2.8
imagezmq==1.1.1
importlib-metadata==1.6.0
imutils==0.5.3
ipaddr==2.2.0
jdcal==1.4.1
lockfile==0.12.2
Markdown==3.1.1
msgpack==0.6.2
mysqlclient==1.4.5
numpy==1.18.4
opencv-python==4.2.0.34
openpyxl==2.6.0
packaging==20.3
pep517==0.8.2
Pillow==7.1.2
pkg-resources==0.0.0
progress==1.5
pycodestyle==2.5.0
pycparser==2.19
pyparsing==2.4.6
Pyphen==0.9.5
python-decouple==3.3
pytoml==0.1.21
pytz==2019.3
pyzmq==19.0.1
requests==2.22.0
retrying==1.3.3
six==1.14.0
sqlparse==0.3.0
tinycss2==1.0.2
toml==0.10.1
urllib3==1.25.7
WeasyPrint==51
webencodings==0.5.1
xlwt==1.3.0
zipp==1.2.0
zmq==0.0.0
